상세 페이지 편집 기능 추가 및 상태 관리 개선, 새로운 입력 필드 및 요약 박스 추가
