+++ app/main/detail/layout.tsx
@@ -1,12 +1,25 @@
 "use client";
+import classNames from "classnames";
+import { format } from "date-fns";
 import { useRouter } from "next/navigation";
-import styles from "../main.module.css";
-import DetailPage from "./page";
-import BackIcon from "@/public/icon_back.svg";
-import ExitIcon from "@/public/icon_exit.svg";
-import { useCurrentRequestState } from "@/store/detail.store";
 
-export default function DetailLayout() {
+import BackIcon from "@/public/icon_back.svg";
+import EstimatedIcon from "@/public/icon_estimated.svg";
+import ExitIcon from "@/public/icon_exit.svg";
+import ImageIcon from "@/public/icon_image.svg";
+import LocationIcon from "@/public/icon_location.svg";
+import ShopIcon from "@/public/icon_shop.svg";
+import TimeIcon from "@/public/icon_time.svg";
+import { useCurrentRequestState } from "@/store/detail.store";
+import { DamageRequestProps } from "@/util/typeDefs";
+
+import styles from "../main.module.css";
+
+export default function DetailLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
   const router = useRouter();
   const currentRequestStore = useCurrentRequestState();
 
@@ -39,7 +52,45 @@ export default function DetailLayout() {
         type={currentRequestStore.current.state}
         ctNumber={currentRequestStore.current.containerNumber}
       />
-      <DetailPage />
+      {children}
     </main>
   );
 }
+
+export const SummaryBox = ({ data }: { data: DamageRequestProps }) => {
+  return (
+    <div className={classNames(styles.box, styles.shadow, styles.summary)}>
+      <span className={styles.label}>
+        <ShopIcon />
+        MNR Vendor Shop
+      </span>
+      <span className={styles.value}>{data.vendorShop || "-"}</span>
+      <span className={styles.label}>
+        <EstimatedIcon />
+        Estimated
+      </span>
+      <span className={styles.value}>{data.estimatedPeriod || "-"}</span>
+      <span className={styles.label}>
+        <LocationIcon />
+        Location
+      </span>
+      <span className={styles.value}>
+        {data.location.latitude}, {data.location.longitude}
+      </span>
+      <span className={styles.label}>
+        <TimeIcon />
+        Time
+      </span>
+      <span className={styles.value}>
+        {format(data.createdAt, "dd/MM/yyyy | HH:mm:ss")}
+      </span>
+      <div className={styles.label}>
+        <ImageIcon height="20" />
+        Images
+      </div>
+      <span className={styles.value}>
+        {`${data.images.length} images added`}
+      </span>
+    </div>
+  );
+};
